+++ packages/arui-scripts/bin/build.sh
@@ -1,0 +1,1 @@
+tsc --project tsconfig-local.json;cp ./src/configs/mq.css build/configs/mq.css;